% Auto-generated: do not edit by hand
\name{dashCharty}

\alias{dashCharty}

\title{DashCharty component}

\description{

}

\usage{
dashCharty(id=NULL, animated=NULL, autoScale=NULL,
buttonTextColor=NULL, colors=NULL, data=NULL, disabled=NULL,
endX=NULL, fillColors=NULL, hideFromLegend=NULL,
legendPosition=NULL, maxY=NULL, minY=NULL, names=NULL,
onZoomIn=NULL, rangeTextType=NULL, showBrush=NULL,
showButtons=NULL, showLegend=NULL, showLegendTitle=NULL,
showMainArea=NULL, showPreview=NULL, showRangeText=NULL,
startX=NULL, stepX=NULL, theme=NULL, themeIdx=NULL,
themes=NULL, title=NULL, type=NULL, xAxisStep=NULL,
xAxisType=NULL, yAxisType=NULL, zoomInterval=NULL,
zoomStepX=NULL)
}

\arguments{
\item{id}{Character. }

\item{animated}{Logical. }

\item{autoScale}{Logical. }

\item{buttonTextColor}{Named list. }

\item{colors}{Named list. }

\item{data}{Named list. }

\item{disabled}{Named list. }

\item{endX}{Numeric. }

\item{fillColors}{Named list. }

\item{hideFromLegend}{Named list. }

\item{legendPosition}{Character. }

\item{maxY}{Numeric. }

\item{minY}{Numeric. }

\item{names}{Named list. }

\item{onZoomIn}{}

\item{rangeTextType}{Character. }

\item{showBrush}{Logical. }

\item{showButtons}{Logical. }

\item{showLegend}{Logical. }

\item{showLegendTitle}{Logical. }

\item{showMainArea}{Logical. }

\item{showPreview}{Logical. }

\item{showRangeText}{Logical. }

\item{startX}{Numeric. }

\item{stepX}{Numeric. }

\item{theme}{Named list. }

\item{themeIdx}{Numeric. }

\item{themes}{Unnamed list. }

\item{title}{Character. }

\item{type}{Character. }

\item{xAxisStep}{Numeric. }

\item{xAxisType}{Character. }

\item{yAxisType}{Character. }

\item{zoomInterval}{Numeric. }

\item{zoomStepX}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

